Refactor storage size fields from BigInt to Int for improved compatibility
- Updated user and track models to change storage-related fields from BigInt to Int, ensuring consistency across the application.
- Modified related functions and middleware to accommodate the new data types.
- Created database migrations to alter existing columns in the database schema, preventing potential data loss during the transition.
- Enhanced error handling and logging in the track conversion process to manage storage quota checks effectively.

+++ packages/core-storage/prisma/seeds/system-settings.ts
@@ -11,7 +11,7 @@ const DEFAULT_SETTINGS = [
   {
     key: SystemSettingKey.FREE_STORAGE_QUOTA_BYTES,
     valueType: SystemSettingValueType.NUMBER,
-    numberValue: BigInt(1024 * 1024 * 1024), // 1GB in bytes
+    numberValue: 1024 * 1024 * 1024, // 1GB in bytes
     description: "Default free storage quota for new users",
   },
 ] as const;
